2272 - Update based on PR comments
