+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/StreamAttachmentFactories.kt
@@ -3,24 +3,29 @@ package io.getstream.chat.android.compose.ui.attachments
 import androidx.compose.runtime.Composable
 import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.compose.state.messages.attachments.AttachmentState
+import io.getstream.chat.android.compose.ui.attachments.factory.FileAttachmentFactory
+import io.getstream.chat.android.compose.ui.attachments.factory.GiphyAttachmentFactory
+import io.getstream.chat.android.compose.ui.attachments.factory.ImageAttachmentFactory
+import io.getstream.chat.android.compose.ui.attachments.factory.LinkAttachmentFactory
 
 public object StreamAttachmentFactories {
 
     /**
-     * The default description length of link attachments. We limit this, because for some links the description
-     * can prove to be too long.
-     * */
-    private const val DEFAULT_LINK_DESCRIPTION_LINE_LIMIT = 5
+     * The default max length of the link attachments description. We limit this, because for some links the description
+     * can be too long.
+     */
+    private const val DEFAULT_LINK_DESCRIPTION_MAX_LINES = 5
 
     /**
      * Default attachment factories we provide, which can transform image, file and link attachments.
      *
-     * @param linkDescriptionLineLimit - The limit of how long the link attachment descriptions can be.
-     * */
+     * @param linkDescriptionMaxLines - The limit of how long the link attachment descriptions can be.
+     * @return A [List] of various [AttachmentFactory] instances that provide different attachments support.
+     */
     public fun defaultFactories(
-        linkDescriptionLineLimit: Int = DEFAULT_LINK_DESCRIPTION_LINE_LIMIT,
+        linkDescriptionMaxLines: Int = DEFAULT_LINK_DESCRIPTION_MAX_LINES,
     ): List<AttachmentFactory> = listOf(
-        LinkAttachmentFactory(linkDescriptionLineLimit),
+        LinkAttachmentFactory(linkDescriptionMaxLines),
         GiphyAttachmentFactory(),
         ImageAttachmentFactory(),
         FileAttachmentFactory()
@@ -33,7 +38,7 @@ public object StreamAttachmentFactories {
  * @param canHandle - Checks the message and returns if the factory can consume it or not.
  * @param content - Composable function that allows users to define the content the [AttachmentFactory] will build using any given
  * [AttachmentState].
- * */
+ */
 public abstract class AttachmentFactory(
     public val canHandle: (attachments: List<Attachment>) -> Boolean,
     public val content: @Composable (AttachmentState) -> Unit,
