+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/content/FileAttachmentContent.kt
@@ -1,4 +1,4 @@
-package io.getstream.chat.android.compose.ui.attachments
+package io.getstream.chat.android.compose.ui.attachments.content
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
@@ -29,15 +29,6 @@ import io.getstream.chat.android.compose.state.messages.attachments.AttachmentSt
 import io.getstream.chat.android.compose.ui.attachments.components.FileAttachmentImage
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 import io.getstream.chat.android.offline.ChatDomain
-
-/**
- * An extension of the [AttachmentFactory] that validates attachments as files and uses [FileAttachmentContent] to
- * build the UI for the message.
- * */
-public class FileAttachmentFactory : AttachmentFactory(
-    canHandle = { attachments -> attachments.isNotEmpty() },
-    content = @Composable { FileAttachmentContent(it) }
-)
 
 /**
  * Builds a file attachment message.
