+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/theme/StreamShapes.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.unit.dp
  * @param imageThumbnail - The shape of image thumbnails, shown in selected attachments and image file attachments.
  * @param bottomSheet - The shape of components used as bottom sheets.
  * */
-public class StreamShapes(
+public data class StreamShapes(
     public val avatar: Shape,
     public val myMessageBubble: Shape,
     public val otherMessageBubble: Shape,
@@ -26,7 +26,7 @@ public class StreamShapes(
     public val bottomSheet: Shape,
 ) {
     public companion object {
-        public val default: StreamShapes = StreamShapes(
+        public fun defaultShapes(): StreamShapes = StreamShapes(
             avatar = CircleShape,
             myMessageBubble = RoundedCornerShape(topStart = 16.dp, topEnd = 16.dp, bottomStart = 16.dp),
             otherMessageBubble = RoundedCornerShape(topStart = 16.dp, topEnd = 16.dp, bottomEnd = 16.dp),
